+++ src/rai_core/rai/communication/ros2/__init__.py
@@ -13,25 +13,17 @@
 # limitations under the License.
 
 from .api import (
-    ConfigurableROS2TopicAPI,
-    IROS2Message,
-    ROS2ActionAPI,
-    ROS2ServiceAPI,
-    ROS2TopicAPI,
-    TopicConfig,
+    IROS2Message,  # TODO: IROS2Message should not be a part of the public API
+    TopicConfig,  # TODO: TopicConfig should not be a part of the public API
 )
 from .connectors import ROS2ARIConnector, ROS2HRIConnector
 from .messages import ROS2ARIMessage, ROS2HRIMessage
 
 __all__ = [
-    "ConfigurableROS2TopicAPI",
     "IROS2Message",
     "ROS2ARIConnector",
     "ROS2ARIMessage",
-    "ROS2ActionAPI",
     "ROS2HRIConnector",
     "ROS2HRIMessage",
-    "ROS2ServiceAPI",
-    "ROS2TopicAPI",
     "TopicConfig",
 ]
